+++ packages/client/src/pages/login/types.ts
@@ -1,0 +1,1 @@
+export type AuthData = { login: string; password: string };
